axios res notifications, delete user done

+++ frontend/src/api/user.js
@@ -2,7 +2,6 @@ import { useContext } from "react"
 import useAxios from "../hooks/useAxios"
 import { useMutation, useQuery } from "@tanstack/react-query"
 import AuthContext from "../contexts/AuthContext"
-import { success } from "../utils"
 
 const prefix = '/api/v1'
 
@@ -18,13 +17,9 @@ export const useDeleteUser = () => {
     const api = useAxios()
     const { logoutUser } = useContext(AuthContext)
     return useMutation({
-        mutationFn: (userId) => {
-            api.delete(`${prefix}/users/${userId}`)
-        },
-        onSuccess: (res) => {
-            // logoutUser() 
-            console.log(res, 'success')
-            success(res.message)
+        mutationFn: (userId) => api.delete(`${prefix}/users/${userId}`),
+        onSuccess: () => {
+            logoutUser()
         }
     })
 }
